+++ core/src/main/java/com/starlight/intrepid/ConnectionListener.java
@@ -31,6 +31,7 @@ import com.starlight.intrepid.auth.UserContextInfo;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.net.InetAddress;
+import java.net.SocketAddress;
 
 
 /**
@@ -51,7 +52,7 @@ public interface ConnectionListener {
 	 * @param connection_type_description   An SPI-specific description of the connection
 	 * @param ack_rate_sec                  Rate of method acks in seconds.
 	 */
-	void connectionOpened( @Nonnull InetAddress host, int port,
+	void connectionOpened( @Nonnull SocketAddress socket_address,
 		@Nullable Object attachment, @Nonnull VMID source_vmid, @Nonnull VMID vmid,
 		@Nullable UserContextInfo user_context, @Nullable VMID previous_vmid,
 		@Nonnull Object connection_type_description, byte ack_rate_sec );
@@ -66,7 +67,7 @@ public interface ConnectionListener {
 	 * @param will_attempt_reconnect        Indicates whether or not a reconnection
 	 * @param user_context                  User context, if any.
 	 */
-	void connectionClosed( @Nonnull InetAddress host, int port,
+	void connectionClosed( @Nonnull SocketAddress socket_address,
 		@Nonnull VMID source_vmid, @Nullable VMID vmid, @Nullable Object attachment,
 		boolean will_attempt_reconnect, @Nullable UserContextInfo user_context );
 
@@ -77,14 +78,12 @@ public interface ConnectionListener {
 	 *
 	 * @param connection_type_description   An SPI-specific description of the connection
 	 */
-	void connectionOpening( @Nonnull InetAddress host,
-		int port, @Nullable Object attachment, @Nullable ConnectionArgs args,
-		@Nonnull Object connection_type_description );
+	void connectionOpening( @Nonnull SocketAddress socket_address, @Nullable Object attachment,
+		@Nullable ConnectionArgs args, @Nonnull Object connection_type_description );
 
 	/**
 	 * Called after {@link #connectionOpening} if the connection could not be made.
 	 */
-	void connectionOpenFailed( @Nonnull InetAddress host,
-		int port, @Nullable Object attachment, @Nullable Exception error,
-		boolean will_retry );
+	void connectionOpenFailed( @Nonnull SocketAddress socket_address, @Nullable Object attachment,
+		@Nullable Exception error, boolean will_retry );
 }