+++ driver/mina/src/test/java/com/starlight/intrepid/TestReconnecting.java
@@ -7,6 +7,7 @@ import com.starlight.intrepid.auth.UserContextInfo;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.net.InetAddress;
+import java.net.SocketAddress;
 import java.util.concurrent.TimeUnit;
 
 
@@ -32,37 +33,37 @@ public class TestReconnecting {
 
 		instance.addConnectionListener( new ConnectionListener() {
 			@Override
-			public void connectionOpened( @Nonnull InetAddress host, int port,
+			public void connectionOpened( @Nonnull SocketAddress socket_address,
 				Object attachment, @Nonnull VMID source_vmid, @Nonnull VMID vmid,
 				UserContextInfo user_context, VMID previous_vmid,
 				@Nonnull Object connection_type_description, byte ack_rate_sec ) {
 
-				System.out.println( "connectionOpened: " + host + ":" + port );
+				System.out.println( "connectionOpened: " + socket_address );
 			}
 
 			@Override
-			public void connectionClosed( @Nonnull InetAddress host, int port,
+			public void connectionClosed( @Nonnull SocketAddress socket_address,
 				@Nonnull VMID source_vmid, @Nullable VMID vmid,
 				@Nullable Object attachment, boolean will_attempt_reconnect,
 				@Nullable UserContextInfo user_context ) {
 
-				System.out.println( "connectionClosed: " + host + ":" + port +
+				System.out.println( "connectionClosed: " + socket_address +
 					( will_attempt_reconnect ? " will reconnect" : " will NOT reconnect" ) );
 			}
 
 			@Override
-			public void connectionOpening( @Nonnull InetAddress host, int port,
+			public void connectionOpening( @Nonnull SocketAddress socket_address,
 				Object attachment, ConnectionArgs args,
 				@Nonnull Object connection_type_description ) {
 
-				System.out.println( "connectionOpening: " + host + ":" + port );
+				System.out.println( "connectionOpening: " + socket_address );
 			}
 
 			@Override
-			public void connectionOpenFailed( @Nonnull InetAddress host, int port,
+			public void connectionOpenFailed( @Nonnull SocketAddress socket_address,
 				Object attachment, Exception error, boolean will_retry ) {
 
-				System.out.println( "connectionOpenFailed: " + host + ":" + port +
+				System.out.println( "connectionOpenFailed: " + socket_address +
 					( will_retry ? " will retry" : " will NOT retry" ) );
 			}
 		} );
@@ -82,31 +83,31 @@ public class TestReconnecting {
 
 		intrepid.addConnectionListener( new ConnectionListener() {
 			@Override
-			public void connectionOpened( @Nonnull InetAddress host, int port,
+			public void connectionOpened( @Nonnull SocketAddress socket_address,
 				Object attachment, @Nonnull VMID source_vmid, @Nonnull VMID vmid,
 				UserContextInfo user_context, VMID previous_vmid,
 				@Nonnull Object connection_type_description, byte ack_rate_sec ) {
 
-				System.out.println( "connectionOpened: " + host + ":" + port );
+				System.out.println( "connectionOpened: " + socket_address );
 			}
 
 			@Override
-			public void connectionClosed( @Nonnull InetAddress host, int port,
+			public void connectionClosed( @Nonnull SocketAddress socket_address,
 				@Nonnull VMID source_vmid, @Nullable VMID vmid,
 				@Nullable Object attachment, boolean will_attempt_reconnect,
 				@Nullable UserContextInfo user_context ) {
 
-				System.out.println( "connectionClosed: " + host + ":" + port +
+				System.out.println( "connectionClosed: " + socket_address +
 					( will_attempt_reconnect ? " will reconnect" : " will NOT reconnect" ) );
 			}
 
 			@Override
-			public void connectionOpening( @Nonnull InetAddress host, int port,
+			public void connectionOpening( @Nonnull SocketAddress socket_address,
 				Object attachment, ConnectionArgs args,
 				@Nonnull Object connection_type_description ) {}
 
 			@Override
-			public void connectionOpenFailed( @Nonnull InetAddress host, int port,
+			public void connectionOpenFailed( @Nonnull SocketAddress socket_address,
 				Object attachment, Exception error, boolean will_retry ) {}
 		} );