+++ driver/mina/src/test/java/com/starlight/intrepid/demo/basic/client/TestClient.java
@@ -36,6 +36,7 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.lang.reflect.Method;
 import java.net.InetAddress;
+import java.net.SocketAddress;
 import java.util.concurrent.TimeUnit;
 
 
@@ -97,7 +98,7 @@ public class TestClient implements ClientInterface {
 		} );
 		intrepid.addConnectionListener( new ConnectionListener() {
 			@Override
-			public void connectionOpened( @Nonnull InetAddress host, int port,
+			public void connectionOpened( @Nonnull SocketAddress socket_address,
 				Object attachment, @Nonnull VMID source_vmid, @Nonnull VMID vmid,
 				UserContextInfo user_context, VMID previous_vmid,
 				@Nonnull Object connection_type_description, byte ack_rate_sec ) {
@@ -107,7 +108,7 @@ public class TestClient implements ClientInterface {
 			}
 
 			@Override
-			public void connectionClosed( @Nonnull InetAddress host, int port,
+			public void connectionClosed( @Nonnull SocketAddress socket_address,
 				@Nonnull VMID source_vmid, @Nullable VMID vmid,
 				@Nullable Object attachment, boolean will_attempt_reconnect,
 				@Nullable UserContextInfo user_context ) {
@@ -118,11 +119,11 @@ public class TestClient implements ClientInterface {
 			}
 
 			@Override
-			public void connectionOpenFailed( @Nonnull InetAddress host, int port,
+			public void connectionOpenFailed( @Nonnull SocketAddress socket_address,
 				Object attachment, Exception error, boolean will_retry ) {}
 
 			@Override
-			public void connectionOpening( @Nonnull InetAddress host, int port,
+			public void connectionOpening( @Nonnull SocketAddress socket_address,
 				Object attachment, ConnectionArgs args,
 				@Nonnull Object connection_type_description ) {}
 		} );